requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.2
opencv-contrib-python==3.4.0.14
opencv-python==3.4.0.12
pprint==0.1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
six==1.11.0
